+++ test/test_sdam_monitoring_spec.py
@@ -197,7 +197,7 @@ def create_test(scenario_def):
                     source_address = clean_node(source)
                     topology.on_change(ServerDescription(
                         address=source_address,
-                        ismaster=Hello(response),
+                        hello=Hello(response),
                         round_trip_time=0))
 
                 expected_results = phase['outcome']['events']
@@ -326,10 +326,10 @@ class TestSdamMonitoring(IntegrationTest):
     def test_network_error_publishes_events(self):
         self._test_app_error({'closeConnection': True}, ConnectionFailure)
 
-    # In 4.4+, NotMaster errors from failCommand don't cause SDAM state
+    # In 4.4+, not primary errors from failCommand don't cause SDAM state
     # changes because topologyVersion is not incremented.
     @client_context.require_version_max(4, 3)
-    def test_not_master_error_publishes_events(self):
+    def test_not_primary_error_publishes_events(self):
         self._test_app_error({'errorCode': 10107, 'closeConnection': False,
                               'errorLabels': ['RetryableWriteError']},
                              NotPrimaryError)