+++ test/utils_selection_tests.py
@@ -23,7 +23,7 @@ sys.path[0:0] = [""]
 from bson import json_util
 from pymongo.common import clean_node, HEARTBEAT_FREQUENCY
 from pymongo.errors import AutoReconnect, ConfigurationError
-from pymongo.hello import Hello
+from pymongo.hello import Hello, HelloCompat
 from pymongo.server_description import ServerDescription
 from pymongo.settings import TopologySettings
 from pymongo.server_selectors import writable_server_selector
@@ -62,30 +62,30 @@ def make_server_description(server, hosts):
     if server_type in ("Unknown", "PossiblePrimary"):
         return ServerDescription(clean_node(server['address']), Hello({}))
 
-    ismaster_response = {'ok': True, 'hosts': hosts}
+    hello_response = {'ok': True, 'hosts': hosts}
     if server_type != "Standalone" and server_type != "Mongos":
-        ismaster_response['setName'] = "rs"
+        hello_response['setName'] = "rs"
 
     if server_type == "RSPrimary":
-        ismaster_response['ismaster'] = True
+        hello_response[HelloCompat.LEGACY_CMD] = True
     elif server_type == "RSSecondary":
-        ismaster_response['secondary'] = True
+        hello_response['secondary'] = True
     elif server_type == "Mongos":
-        ismaster_response['msg'] = 'isdbgrid'
+        hello_response['msg'] = 'isdbgrid'
 
-    ismaster_response['lastWrite'] = {
+    hello_response['lastWrite'] = {
         'lastWriteDate': make_last_write_date(server)
     }
 
     for field in 'maxWireVersion', 'tags', 'idleWritePeriodMillis':
         if field in server:
-            ismaster_response[field] = server[field]
+            hello_response[field] = server[field]
 
-    ismaster_response.setdefault('maxWireVersion', 6)
+    hello_response.setdefault('maxWireVersion', 6)
 
     # Sets _last_update_time to now.
     sd = ServerDescription(clean_node(server['address']),
-                           Hello(ismaster_response),
+                           Hello(hello_response),
                            round_trip_time=server['avg_rtt_ms'] / 1000.0)
 
     if 'lastUpdateTime' in server: